Attempt to fix style.css link
The home page is not displaying any of the style changes from the style.css file, even though the syntax for both selectors and the css file is correct. This commit is an attempt to fix the link problem by reorganizing the link tag to be below the title tag at the head of the file.